Removed a x64 compiler warning.

+++ Source/Platforms/WinINET.cpp
@@ -174,7 +174,7 @@ namespace Wininet
         HTTPRequest += "\r\n";
 
         // Send to winsock that forwards it to the server.
-        send(Activerequests[hRequest].Socket, HTTPRequest.c_str(), HTTPRequest.size(), NULL);
+        send(Activerequests[hRequest].Socket, HTTPRequest.c_str(), (int)HTTPRequest.size(), NULL);
 
         return TRUE;
     }
@@ -200,7 +200,7 @@ namespace Wininet
         HTTPRequest += "\r\n";
 
         // Send to winsock that forwards it to the server.
-        send(Activerequests[hRequest].Socket, HTTPRequest.c_str(), HTTPRequest.size(), NULL);
+        send(Activerequests[hRequest].Socket, HTTPRequest.c_str(), (int)HTTPRequest.size(), NULL);
 
         return TRUE;
     }
@@ -225,7 +225,7 @@ namespace Wininet
             HTTPRequest.append((char *)lpOptional, dwOptionalLength);
 
         // Send to winsock that forwards it to the server.
-        send(Activerequests[hRequest].Socket, HTTPRequest.data(), HTTPRequest.size(), NULL);
+        send(Activerequests[hRequest].Socket, HTTPRequest.data(), (int)HTTPRequest.size(), NULL);
 
         return TRUE;
     }
@@ -236,7 +236,7 @@ namespace Wininet
         std::wstring Temporary = lpszHeaders;
         std::string Headers = { Temporary.begin(), Temporary.end() };
 
-        return HTTPSendrequestA(hRequest, Headers.c_str(), Headers.size(), lpOptional, dwOptionalLength);
+        return HTTPSendrequestA(hRequest, Headers.c_str(), (int)Headers.size(), lpOptional, dwOptionalLength);
     }
 
     BOOL __stdcall InternetQueryoptionA(const size_t hInternet, DWORD dwOption, LPVOID lpBuffer, LPDWORD lpdwBufferLength)
